some basic fixes
